Create StripeCheckout
This program uses the Stripe API to create a checkout session for a single product. Here's how the code works:

It defines a function called create_checkout_session() that takes two arguments: the name of the product and the price of the product.
We create a new checkout session using stripe.checkout.Session.create() with the following parameters:
payment_method_types: The payment methods accepted by the checkout session. In this case, we're accepting credit cards.
line_items: A list of line items representing the product being purchased. Each line item contains information about the product name, price, and quantity.
mode: The mode of the checkout session. In this case, we're creating a payment session.
success_url: The URL to redirect the customer to when the payment is successful.
cancel_url: The URL to redirect the customer to when the payment is canceled or fails.

We return the ID of the checkout session using session.id.

To use it, you'll need your Stripe API, but then you  call the create_checkout_session() function with the name and price of the product you want to sell. This will return the ID of the checkout session, which you can use to redirect the customer to the Stripe checkout page. Note that you'll need to replace "YOUR_SECRET_KEY_HERE" with your actual Stripe API key.